refactor(desktop): co-locate TopBar with dashboard layout
Move TopBar from screens/main/components to routes/_authenticated/_dashboard/components
to properly co-locate it next to the layout that uses it. Update import paths accordingly.
