+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/components/TopBar/TopBar.tsx
@@ -1,9 +1,9 @@
 import { useParams } from "@tanstack/react-router";
 import { electronTrpc } from "renderer/lib/electron-trpc";
-import { OpenInMenuButton } from "./OpenInMenuButton";
-import { OrganizationDropdown } from "./OrganizationDropdown";
-import { SidebarToggle } from "./SidebarToggle";
-import { WindowControls } from "./WindowControls";
+import { OpenInMenuButton } from "./components/OpenInMenuButton";
+import { OrganizationDropdown } from "./components/OrganizationDropdown";
+import { SidebarToggle } from "./components/SidebarToggle";
+import { WindowControls } from "./components/WindowControls";
 
 export function TopBar() {
 	const { data: platform } = electronTrpc.window.getPlatform.useQuery();
