+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/components/TopBar/components/WindowControls/index.ts
