+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/components/TopBar/index.ts
